ESP-API/pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["NeZorinEgor <zorin-ep@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pydantic-settings = "^2.2.1"
uvicorn = "^0.28.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
pymysql = "^1.1.0"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
